+++ event-store-management/src/test/java/uk/gov/justice/services/eventstore/management/rebuild/commands/RebuildCommandHandlerTest.java
@@ -1,0 +1,42 @@
+package uk.gov.justice.services.eventstore.management.rebuild.commands;
+
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
+
+import uk.gov.justice.services.common.util.UtcClock;
+import uk.gov.justice.services.eventstore.management.rebuild.events.RebuildRequestedEvent;
+
+import java.time.ZonedDateTime;
+
+import javax.enterprise.event.Event;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.runners.MockitoJUnitRunner;
+
+@RunWith(MockitoJUnitRunner.class)
+public class RebuildCommandHandlerTest {
+
+    @Mock
+    private UtcClock clock;
+
+    @Mock
+    private Event<RebuildRequestedEvent> rebuildRequestedEventEventFirer;
+
+    @InjectMocks
+    private RebuildCommandHandler rebuildCommandHandler;
+
+    @Test
+    public void shouldFireRebuildEvent() throws Exception {
+
+        final ZonedDateTime now = new UtcClock().now();
+
+        when(clock.now()).thenReturn(now);
+
+        rebuildCommandHandler.doRebuild();
+
+        verify(rebuildRequestedEventEventFirer).fire(new RebuildRequestedEvent(now, new RebuildCommand()));
+    }
+}