+++ event-store-util/src/test/java/uk/gov/justice/services/eventsourcing/util/jee/timer/TimerServiceManagerTest.java
@@ -1,13 +1,9 @@
 package uk.gov.justice.services.eventsourcing.util.jee.timer;
 
-import static java.util.Arrays.asList;
 import static org.mockito.Mockito.inOrder;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.verifyZeroInteractions;
 import static org.mockito.Mockito.when;
 
-import javax.ejb.Timer;
 import javax.ejb.TimerConfig;
 import javax.ejb.TimerService;